+++ openvpn/client/optfilt.hpp
@@ -25,6 +25,7 @@
 #include <vector>
 
 #include <openvpn/common/options.hpp>
+#include <openvpn/common/numeric_cast.hpp>
 #include <openvpn/common/string.hpp>
 #include <openvpn/client/dns.hpp>
 
@@ -136,7 +137,7 @@ class PushedOptionsFilter : public OptionList::FilterBase
         if (pushed.size() < match.size())
             return false;
 
-        int i = match.size() - 1;
+        int i = numeric_cast<int>(match.size() - 1);
         if (!string::starts_with(pushed.get(i, -1), match.get(i, -1)))
             return false;
