+++ openvpn/transport/client/transbase.hpp
@@ -52,7 +52,7 @@ struct TransportClient : public virtual RC<thread_unsafe_refcount>
     virtual bool transport_send_queue_empty() = 0;
     virtual bool transport_has_send_queue() = 0;
     virtual void transport_stop_requeueing() = 0;
-    virtual unsigned int transport_send_queue_size() = 0;
+    virtual size_t transport_send_queue_size() = 0;
     virtual void reset_align_adjust(const size_t align_adjust) = 0;
     virtual IP::Addr server_endpoint_addr() const = 0;
     virtual unsigned short server_endpoint_port() const
